+++ src/components/Users/Details/DeleteUser.tsx
@@ -1,11 +1,11 @@
 'use client'
 
 import { deleteUser } from "@/services/Users";
-import { IconButton } from "@chakra-ui/react";
 import { toaster } from "@/components/ui/toaster";
 import { useQueryClient } from "@tanstack/react-query";
 import React from "react";
-import { FaTimesCircle } from "react-icons/fa";
+import { FaTrash } from "react-icons/fa";
+import StyledIconButton from "@/components/Generic/StyledIconButton";
 
 type deleteUserProps = {
   id: string;
@@ -14,29 +14,32 @@ const DeleteUsers = ({ id }: deleteUserProps) => {
   const queryClient = useQueryClient();
 
   const deleteUserDetail = async (id: string) => {
+    if (!confirm("Are you sure you want to delete this user?")) {
+      return;
+    }
+    
     await deleteUser(id);
 
     queryClient.refetchQueries({ queryKey: ["users"] });
 
     toaster.create({
       title: "Success",
-      description: "User deleted",
+      description: "User deleted successfully",
       type: "success",
       duration: 5000,
     });
   };
 
   return (
-    <IconButton
-      size="sm"
-      cursor="pointer"
-      color="red.500"
-      as={FaTimesCircle}
-      aria-label="delete product"
+    <StyledIconButton
+      variant="delete"
+      aria-label="Delete user"
       onClick={async () => {
         await deleteUserDetail(id.toString());
       }}
-    />
+    >
+      <FaTrash />
+    </StyledIconButton>
   );
 };
